вывел пароль в отдельный класс
